requirements.txt
asgiref==3.11.1
Django==5.2.11
gunicorn==25.1.0
numpy==1.26.4
packaging==26.0
pillow==12.1.1
sqlparse==0.5.5
tzdata==2025.3
